requirements.txt
requests
jinja2
playwright
Pillow